pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "compose-rename"
version = "0.3.1"
description = "Rename a Docker Compose project by migrating volumes to a new project prefix."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
  { name = "Your Name" }
]
dependencies = [
  "pyyaml>=6.0"
]

[project.urls]
Homepage = "https://github.com/jonasjancarik/compose-rename"

[project.scripts]
compose-rename = "compose_rename:main"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
    "pytest-timeout>=2.4.0",
]


